requirements.txt
fastapi>=0.115.12
uvicorn[standard]>=0.34.2
asyncpg>=0.30.0
python-dotenv>=1.0.0
bcrypt>=4.2.2
pydantic[email]>=2.10.6
python-jose[cryptography]>=3.3.0
chromadb>=0.6.5
requests>=2.32.3
authlib>=1.3.0
httpx>=0.27.0
itsdangerous>=2.1.2
